Fix issue where test failures not reported
Seems like getting a hard exit(crash) when a helper exists that didn't
have a `_failed` function. So adding this to the base class.

This was causing the following parts of the promise chain to not get
called, which surfaced itself as the tests stopping with no errors being
reported (and the browser window being left open)

Ref: https://github.com/Codeception/CodeceptJS/issues/93