Add seeNumberOfVisibleElements
Give the ability to not only see how many times an element appears in the DOM but how many times an element is actually visible.

Had to build this function as a custom helper for my project. I believe this should be implemented. Very useful, especially when writing acceptance tests.