fix references to support objects when using DI (#1701)
some change to how the support objects are initialized removed the original
reference to the support object and replaced it a "cloned" object. However,
since Object.assign was used, not all properties have been copied and especially
the original reference was lost.

This implementation uses a proxy object to load the injected modules on demand
and returns the original reference to the module. This should fix all usages
of support objects, where the type is not a plain object.