fix missing support object definition in steps.d.ts
due to support object being a Proxy now, a few more traps need
to be implemented to allow more operations on the support object.

Unfortunately I could not find a way to make this compatible
to for-in loops, so when looping of the support object, the
keys need to be fetched explicitly.