Create Puppeteer helper using container
In order for the session() method to work a couple of things
needed to happen.

First we needed to start the recorder
so its isRunning property would be set to true and the functions
passed in session to recorder would actually run.

Secondly we in order for a session to be successfully created
we need to create the new Puppeteer helper using the container create
method this ensures the call to container.helpers() in session
will actually return something.

Finally we assign the constructed Puppeteer helper to I
so the tests still work properly