Align output from run-workers
This change casts the output process number into a String and checks for the
string length. If it consists of a single character, the `outputProcess` value
is padded with a `0`. This preserves even output when running with more than 9
workers.

This makes it substantially easier to visually parse the test output when using
larger numbers of workers, for example 30.