Fix: allure attachment fails if screenshot failed
When `helper.saveScreenshot` fails, we can stop the execution of this
function.

When the screenshot is not there and the allure report is enabled, it
will throw the folling error:
```
       "after each" hook: codeceptjs.after for "...":
     Uncaught ENOENT: no such file or directory, open '/code/output/record_19bee45168cf2898aee9250e1bbb630e/0024.png'
      at Object.openSync (node:fs:585:3)
      at Object.readFileSync (node:fs:453:35)
      at EventEmitter.persistStep (node_modules/codeceptjs/lib/plugin/stepByStepReport.js:165:69)
```

So if there is an error with the screenshot, we will not add it to the
list of artifacts, and will not try to added it to the allure report.