fix: don't report files as existing when non-directory is in path
fileExists would return true when testing "somedir/somefile" if
somedir existed but was a file, not a directory. This was caused
by checking for ENOENT only, when in that case the error code
returned by "statSync" is ENOTDIR.

Replacing the use of "statSync" by a call to "fs.existsSync"
avoids the problem while working correctly for all other cases.