JsDoc: Remove promise from `Actor.say`
The return type of `Actor.say` contains `Promise<any>`. In TypeScript code, with the ESLint rule "@typescript-eslint/no-floating-promises" this causes lint errors for every `I.say` in the tests stating that a promise will not be awaited.
Instead, add a comment similar to many other methods (`I.see`, `I.click`, ...) and leave the return type `void`.